Cart Service: Proof of concept with QRs
This is the first POC for the cart service. The idea is to validate
how the service will work, from the vendor point of view.

A single index.html has been added, where items can be added to a cart
for now, only the ID without checking nor persisting anything.

The page allows 2 URLs, the main allowing to add elements to carts
and the other adding direct element to a cart, all the carts have
a session ID that depends on the cross origin request, so if the vendor
scans multiple QRs from its device or mobilePhone, they will be added
to the cart automatically, requiring nothing else than conectivity
to the main page

Signed-off-by: OscarMoya <oscar.moya@nearbycomputing.com>